+++ src/axom/core/tests/core_Path.hpp
@@ -14,6 +14,13 @@ TEST(core_Path, trivial_construct)
   EXPECT_EQ(static_cast<std::string>(path), path_str);
 }
 
+TEST(core_Path, lead_delim_construct)
+{
+  const std::string path_str = "/foo";
+  axom::Path path(path_str);
+  EXPECT_EQ(static_cast<std::string>(path), path_str);
+}
+
 TEST(core_Path, two_component_construct)
 {
   const std::string path_str = "foo/bar";
@@ -25,6 +32,27 @@ TEST(core_Path, two_component_construct_alt_delim)
 {
   const std::string path_str = "foo:bar";
   axom::Path path(path_str, ':');
+  EXPECT_EQ(static_cast<std::string>(path), path_str);
+}
+
+TEST(core_Path, two_component_construct_lead_delim)
+{
+  const std::string path_str = "/foo/bar";
+  axom::Path path(path_str);
+  EXPECT_EQ(static_cast<std::string>(path), path_str);
+}
+
+TEST(core_Path, two_component_construct_alt_lead_delim)
+{
+  const std::string path_str = ":foo:bar";
+  axom::Path path(path_str, ':');
+  EXPECT_EQ(static_cast<std::string>(path), path_str);
+}
+
+TEST(core_Path, no_component_construct_root_delim)
+{
+  const std::string path_str = "/";
+  axom::Path path(path_str);
   EXPECT_EQ(static_cast<std::string>(path), path_str);
 }
 
@@ -44,6 +72,22 @@ TEST(core_Path, join_trivial_alt_delim)
   EXPECT_EQ(static_cast<std::string>(joined), "foo:bar");
 }
 
+TEST(core_Path, join_trivial_lead_delim)
+{
+  axom::Path first("/foo");
+  axom::Path second("/bar");
+  auto joined = axom::Path::join({first, second});
+  EXPECT_EQ(static_cast<std::string>(joined), "/foo/bar");
+}
+
+TEST(core_Path, join_trivial_alt_lead_delim)
+{
+  axom::Path first(":foo", ':');
+  axom::Path second(":bar", ':');
+  auto joined = axom::Path::join({first, second}, ':');
+  EXPECT_EQ(static_cast<std::string>(joined), ":foo:bar");
+}
+
 TEST(core_Path, join_multi_component)
 {
   axom::Path first("foo/bar");
@@ -60,6 +104,22 @@ TEST(core_Path, join_multi_component_alt_delim)
   EXPECT_EQ(static_cast<std::string>(joined), "foo:bar:baz:quux");
 }
 
+TEST(core_Path, join_multi_component_lead_delim)
+{
+  axom::Path first("/foo/bar");
+  axom::Path second("/baz/quux");
+  auto joined = axom::Path::join({first, second});
+  EXPECT_EQ(static_cast<std::string>(joined), "/foo/bar/baz/quux");
+}
+
+TEST(core_Path, join_multi_component_alt_lead_delim)
+{
+  axom::Path first(":foo:bar", ':');
+  axom::Path second(":baz:quux", ':');
+  auto joined = axom::Path::join({first, second}, ':');
+  EXPECT_EQ(static_cast<std::string>(joined), ":foo:bar:baz:quux");
+}
+
 TEST(core_Path, parent_basic)
 {
   axom::Path path("foo/bar");
@@ -72,6 +132,27 @@ TEST(core_Path, parent_basic_alt_delim)
   axom::Path path("foo;bar", ';');
   auto parent = path.parent();
   EXPECT_EQ(static_cast<std::string>(parent), "foo");
+}
+
+TEST(core_Path, parent_basic_lead_delim)
+{
+  axom::Path path("/foo/bar");
+  auto parent = path.parent();
+  EXPECT_EQ(static_cast<std::string>(parent), "/foo");
+}
+
+TEST(core_Path, parent_basic_alt_lead_delim)
+{
+  axom::Path path(":foo:bar", ':');
+  auto parent = path.parent();
+  EXPECT_EQ(static_cast<std::string>(parent), ":foo");
+}
+
+TEST(core_Path, parent_basic_root_lead_delim)
+{
+  axom::Path path("/foo");
+  auto parent = path.parent();
+  EXPECT_EQ(static_cast<std::string>(parent), "/");
 }
 
 TEST(core_Path, parent_multi_component)
@@ -88,6 +169,20 @@ TEST(core_Path, parent_multi_component_alt_delim)
   EXPECT_EQ(static_cast<std::string>(parent), "foo;bar");
 }
 
+TEST(core_Path, parent_multi_component_lead_delim)
+{
+  axom::Path path("/foo/bar/baz");
+  auto parent = path.parent();
+  EXPECT_EQ(static_cast<std::string>(parent), "/foo/bar");
+}
+
+TEST(core_Path, parent_multi_component_alt_lead_delim)
+{
+  axom::Path path("|foo|bar|baz", '|');
+  auto parent = path.parent();
+  EXPECT_EQ(static_cast<std::string>(parent), "|foo|bar");
+}
+
 TEST(core_Path, dirname_basename_basic)
 {
   axom::Path path("foo/bar");
@@ -100,6 +195,27 @@ TEST(core_Path, dirname_basename_basic_alt_delim)
   axom::Path path("foo;bar", ';');
   EXPECT_EQ(path.dirName(), "foo");
   EXPECT_EQ(path.baseName(), "bar");
+}
+
+TEST(core_Path, dirname_basename_basic_lead_delim)
+{
+  axom::Path path("/foo/bar");
+  EXPECT_EQ(path.dirName(), "/foo");
+  EXPECT_EQ(path.baseName(), "bar");
+}
+
+TEST(core_Path, dirname_basename_basic_alt_lead_delim)
+{
+  axom::Path path("+foo+bar", '+');
+  EXPECT_EQ(path.dirName(), "+foo");
+  EXPECT_EQ(path.baseName(), "bar");
+}
+
+TEST(core_Path, dirname_basename_root_lead_delim)
+{
+  axom::Path path("/foo");
+  EXPECT_EQ(path.dirName(), "/");
+  EXPECT_EQ(path.baseName(), "foo");
 }
 
 TEST(core_Path, dirname_basename_multi_component)
@@ -116,15 +232,36 @@ TEST(core_Path, dirname_basename_multi_component_alt_delim)
   EXPECT_EQ(path.baseName(), "baz");
 }
 
+TEST(core_Path, dirname_basename_multi_component_lead_delim)
+{
+  axom::Path path("/foo/bar/baz");
+  EXPECT_EQ(path.dirName(), "/foo/bar");
+  EXPECT_EQ(path.baseName(), "baz");
+}
+
+TEST(core_Path, dirname_basename_multi_component_alt_lead_delim)
+{
+  axom::Path path("|foo|bar|baz", '|');
+  EXPECT_EQ(path.dirName(), "|foo|bar");
+  EXPECT_EQ(path.baseName(), "baz");
+}
+
 TEST(core_Path, basic_equals)
 {
   axom::Path first("foo/bar/baz");
   axom::Path second("foo/bar/baz");
   axom::Path third("foo:bar:baz", ':');
+  axom::Path fourth("/foo/bar/baz");
+  axom::Path fifth(":foo:bar:baz", ':');
   EXPECT_TRUE(first == second);
   EXPECT_FALSE(first != second);
   EXPECT_FALSE(first == third);
   EXPECT_TRUE(first != third);
+  EXPECT_FALSE(first == fourth);
+  EXPECT_TRUE(first != fourth);
+  EXPECT_FALSE(first == fifth);
+  EXPECT_FALSE(third == fifth);
+  EXPECT_TRUE(third != fifth);
 }
 
 TEST(core_Path, parts_empty)
@@ -183,9 +320,27 @@ TEST(core_Path, basic_split)
   EXPECT_EQ(pair.second, "baz");
 }
 
+TEST(core_Path, basic_split_lead_delim)
+{
+  axom::Path path("/foo/bar/baz");
+  auto pair = path.split();
+
+  EXPECT_EQ(pair.first, "/foo/bar");
+  EXPECT_EQ(pair.second, "baz");
+}
+
 TEST(core_Path, basic_split_dirname_basename)
 {
   axom::Path path("foo/bar/baz");
+  auto pair = path.split();
+
+  EXPECT_EQ(pair.first, path.dirName());
+  EXPECT_EQ(pair.second, path.baseName());
+}
+
+TEST(core_Path, basic_split_dirname_basename_lead_delim)
+{
+  axom::Path path("/foo/bar/baz");
   auto pair = path.split();
 
   EXPECT_EQ(pair.first, path.dirName());
@@ -210,12 +365,12 @@ TEST(core_Path, split_empty_basename)
   EXPECT_EQ(pair.second, "foo");
 }
 
-TEST(core_Path, split_empty_dirname)
+TEST(core_Path, split_root_dirname)
 {
   axom::Path path("/foo");
   auto pair = path.split();
 
-  EXPECT_EQ(pair.first, "");
+  EXPECT_EQ(pair.first, "/");
   EXPECT_EQ(pair.second, "foo");
 }