+++ src/axom/core/tests/core_about.hpp
@@ -8,20 +8,22 @@
 #include "axom/config.hpp"
 #include "axom/core/utilities/About.hpp"
 
-// C/C++ includes
-#include <sstream>  // for std::ostringstream
+#include <sstream>
+#include <string>
 
 TEST(core_about, print_about) { axom::about(); }
 
 //-----------------------------------------------------------------------------
 TEST(core_about, get_version)
 {
-  std::ostringstream EXPECTED_VERSION_STRING;
-  EXPECTED_VERSION_STRING << AXOM_VERSION_FULL;
-#ifdef AXOM_VERSION_EXTRA
-  EXPECTED_VERSION_STRING << "-" << AXOM_VERSION_EXTRA;
-#endif
+  std::ostringstream expected_version;
+  expected_version << AXOM_VERSION_FULL;
+  std::string sha = axom::gitSHA();
+  if(!sha.empty())
+  {
+    expected_version << "-" << sha;
+  }
 
   std::string axom_version = axom::getVersion();
-  EXPECT_EQ(EXPECTED_VERSION_STRING.str(), axom_version);
+  EXPECT_EQ(expected_version.str(), axom_version);
 }