+++ src/axom/sidre/spio/IOManager.cpp
@@ -796,18 +796,12 @@ std::string IOManager::getFileNameForRank(const std::string& file_pattern,
 
   //If the root file was given as a path,
   //find the directory and add it to file_name
-  std::string delimiter = "/";
+  char delimiter = '/';
   axom::Path root_path(root_name);
   std::string root_dir = root_path.dirName();
 
-  if(!root_name.empty() &&
-     axom::utilities::string::startsWith(root_name, delimiter[0]))
-  {
-    root_dir = delimiter + root_dir;
-  }
-
   file_name =
-    axom::utilities::string::appendPrefix(root_dir, file_name, delimiter[0]);
+    axom::utilities::string::appendPrefix(root_dir, file_name, delimiter);
 
   return file_name;
 }